requirements.txt
Django==3.0.6
channels==2.4.0
psycopg2==2.8.5
channels-redis==2.4.2
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
djangorestframework==3.11.0
